core/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
joblib==1.2.0
numpy==1.24.3
psycopg2-binary==2.9.6
PyJWT==2.6.0
pytz==2023.3
requests==2.29.0
scikit-learn==1.2.2
scipy==1.10.1
sqlparse==0.4.4
threadpoolctl==3.1.0
urllib3==1.26.15
